Feat : 민원서식(ue2) 등록,수정 기능 구현 및 기존 id컬럼 long으로변경

+++ src/main/java/com/example/approval24/dao/AccountDAO.java
@@ -6,6 +6,6 @@ import com.example.approval24.domain.AccountDTO;
 
 @Mapper
 public interface AccountDAO {
-	public AccountDTO findById(int accountId);
+	public AccountDTO findById(long accountId);
 
 }
