+++ src/main/java/com/example/approval24/domain/ComplainRegDTO.java
@@ -9,9 +9,9 @@ import lombok.Data;
 @Data
 public class ComplainRegDTO {
 
-	private int complainCategoryId;
+	private long complainCategoryId;
 	
-	private int complainuserNo;
+	private long complainuserNo;
 
 	private String complainuserResiNoFront; // 주민번호 앞자리(6)
 	private String complainuserResiNoBack; // 주민번호 뒷자리(7)
